feat: Extract out HybridBase/HybridChild
